Sized numbers under one set of operators, bytes as `#[UInt8]`, and `++`
**Sized numeric types.** `Int8`, `Int16`, `Int32`, `UInt8`, `UInt16`,
`UInt32`, `UInt64` and `Float32` join `Int`, `BigInt` and `Float`; `Int64` and
`Float64` are other names for `Int` and `Float`. They are unboxed on all three
engines -- a VM value is still 16 bytes -- in registers, constructor fields and
array slots. Fixed widths wrap, and every conversion (`toInt8` ... `toUInt64`,
`toInt`, `toBigInt`, from a `BigInt` too) keeps the low bits. The arithmetic
is shared in `meadow_core::num`, so the engines cannot drift apart; integers of
different types compare and hash by value.

**One set of operators.** `+ - * / % ^`, the comparisons and the bitwise
primitives work on every integer type, through a number class `n` that a
`fun` generalizes (`fun square x = x * x` is `forall n. n -> n`) and a `def` or
`let` does not. A literal nothing settles is a `BigInt`. The `~` operators
are gone. `+.` and friends take `Float` or `Float32` through a class `f` that
defaults to `Float`. A mismatch reads "`String` is not an integer type".

**`Std.Num`.** `Std.Int` and `Std.Bits` move to `Std.Num.Int` and
`Std.Num.Bits`, with the bounds of each size in `Std.Num`; the bit helpers
work at the width of the value they are given.

**Bytes are `#[UInt8]`**, written with the new `#[T]` type syntax: the bytes
primitives, `Std.Bytes` (now with 64-bit reads and writes), `Std.String`,
`Json`, `Path`, `Fs.readBytes` and the mini-ml parser.

**`++`** is `Std.String.concat`, re-exported by the prelude. It is the first
operator that is a name rather than a primitive: bound with `fun (++) a b`
or `def (++) = ...`, imported with `use M (++)`, shadowed like anything else,
right-associative beside `::`. Names like it print as `(++)`.

**Editor.** Inlay hints and debugger types name variables by kind -- `n` for
an integer type, `e` for an effect, `r` for a row -- as schemes already did, so
a hint is an annotation you could write. Type completion offers the new
types.

Also: an annotated parameter no longer allocates a closure per call, and the
tutorial covers the numeric types, the `#[UInt8]` digit-fold trap and `++`.

+++ buildtools/meadow/tests/bytes.rs
@@ -23,8 +23,8 @@ fn hex() {
 #[test]
 fn hex_prim_types() {
     insta::assert_snapshot!(schemes("def a = bytesToHex\ndef b = bytesFromHex\n"), @r"
-    a : #[Int] -> String
-    b : String -> Maybe #[Int]
+    a : #[UInt8] -> String
+    b : String -> Maybe #[UInt8]
     ");
 }
